backend_Django/Famesta/requirements.txt
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.1.1
